enhanced booking model

+++ Egyptopia/Controllers/RoomController.cs
@@ -1,37 +1,39 @@
-﻿using Egyptopia.Application.Repositories;
+﻿using AutoMapper;
+using Egyptopia.Application.Repositories;
 using Egyptopia.Domain.Entities;
+using EgyptopiaApi.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using System.Collections.Generic;
 
 namespace EgyptopiaApi.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    //[ApiExplorerSettings(GroupName ="Room")]
+
     public class RoomController : ControllerBase
     {
         private readonly IRoomRepository _roomRepository;
+        private readonly IMapper _mapper;
 
-        public RoomController(IRoomRepository roomRepository)
+        public RoomController(
+            IRoomRepository roomRepository,
+            IMapper mapper)
         {
             _roomRepository = roomRepository;
-        }
-
-        [HttpGet(nameof(GetAllRooms))]
-        public ActionResult<Room?> GetAllRooms()
-        {
-            return Ok(_roomRepository.GetAll());
-        }
-
-        [HttpGet(nameof(GetRoom))]
-        public ActionResult<Room?> GetRoom(Guid id)
-        {
-            return Ok(_roomRepository.Get(id));
+            _mapper = mapper;
         }
 
         [HttpPost(nameof(CreateRoom))]
-        public ActionResult<Room?> CreateRoom(Room room)
+        public ActionResult<RoomModel?> CreateRoom(RoomModel model)
         {
-            var data = _roomRepository.Create(room);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var data = _roomRepository.Create(_mapper.Map<Room>(model));
             if (data == null)
             {
                 return BadRequest();
@@ -39,22 +41,45 @@ namespace EgyptopiaApi.Controllers
             return Ok(data);
         }
 
-        [HttpPut(nameof(UpdateRoom))]
-        public ActionResult<Room?> UpdateRoom(Room room)
+        //[Authorize]
+        [HttpGet(nameof(GetAllRoom))]
+        public ActionResult<List<RoomModel>> GetAllRoom()
         {
-            if (room == null)
+            return Ok(_mapper.Map<List<RoomModel>>(_roomRepository.GetAll()));
+        }
+
+        [HttpGet(nameof(GetRoom))]
+        public ActionResult<RoomModel?> GetRoom(Guid id)
+        {
+            return Ok(_mapper.Map<RoomModel>(_roomRepository.Get(id)));
+        }
+
+        [HttpPut(nameof(UpdateRoom))]
+        public ActionResult<RoomModel?> UpdateRoom(RoomModel model)
+        {
+            if (!ModelState.IsValid)
             {
-                return BadRequest();
+                return BadRequest(ModelState);
             }
-            return Ok(_roomRepository.Update(room));
+            if (model == null)
+                return BadRequest();
+            var entity = _roomRepository.Get(model.Id);
+            if (entity == null)
+                return NotFound();
+            return Ok(_roomRepository.Update(_mapper.Map(model, entity)));
         }
 
         [HttpDelete(nameof(DeleteRoom))]
         public ActionResult DeleteRoom(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                return BadRequest("Id cant be empty");
+            }
             var entity = _roomRepository.Get(id);
             if (entity == null)
                 return NotFound();
+
             _roomRepository.Delete(entity);
             return Ok();
         }
